etc/updater/requirements.txt
flask==2.0.1
requests==2.26.0
prometheus-client==0.11.0
python-json-logger==2.0.2
pyyaml==5.4.1
docker==5.0.3
gunicorn==20.1.0
python-dotenv==0.19.0
redis==4.0.2